requirements.txt
pillow
pytesseract\
mss
pyautogui
keyboard
opencv-python
numpy
